requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
dataclasses-json==0.5.7
dnspython==2.3.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
langchain==0.0.181
loguru==0.7.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.7
openapi-schema-pydantic==1.2.4
packaging==23.1
pinecone-client==2.2.1
pydantic==1.10.8
python-dateutil==2.8.2
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.15
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.2
urllib3==2.0.2
win32-setctime==1.1.0
yarl==1.9.2
